+++ examples/HSE/verilua/setvars.sh
@@ -1,3 +1,3 @@
-# Set variables for verilua. Only VERILUA_CFG is required, which is a path to a configuration file(in Lua).
+# Set variables for verilua. Only VL_CFG_FILE is required, which is a path to a configuration file(in Lua).
 # Verilua use the configuration file to setup the runtime environment.
-export VERILUA_CFG="$(pwd)/cfg.lua"
+export VL_CFG_FILE="$(pwd)/cfg.lua"